Make Python file names snake_case
Not only is it annoying to have all of the errors constantly popping up
warning than the file name is not compliant with the Python snake_case
naming style but it's annoying to put in backslashes in front of every
space in the command line. Spaces are also just generally annoying when
it comes to coding file paths because sometimes you need to put in
backslashes in front of the statements and with some tools you don't.
All to say, spaces are annoying in file names and they are now gone.