Add Axis and Allies Battle Calc Script
Added the very first rough draft of a Python script to compute the
winner of a land battle in the board game Axis and Allies. It can only
do a completely land based battle so far. I plan to add functionality to
add other battle types, simulations, and graphs of results.

Signed-off-by: realprogrammersusevim <logouts.deco-0e@icloud.com>