Add Script To Scrape All xkcd Comics
Added a Python script to scrape every single xkcd comic. It begins at
at the xkcd home page and then attempts to find the image and then
download it into a folder. If it can’t find the image then that is
probably because it isn’t the expected .png image but an interactive
program likely running Javascript. If it runs into that error than it
simply skips the comic and goes onto the next page.

Signed-off-by: realprogrammersusevim <logouts.deco-0e@icloud.com>