Add tool to transcribe English into Tengwar
Are you techy? Do you write in Latex? Are you are Lord of the Rings fan?
Then you are apart of an extremely small segment of people who this tool
will actually be useful for. There is a Latex package that translates
Latex commands into Tengwar runes. If you’ve ever tried to use this
you know it is an absolute pain to tranlate everything by hand. This
shell script comes to your rescue by using a Perl script I found as the
backend to translate any string into Tengwar for use in Latex. Have fun!

Signed-off-by: realprogrammersusevim <logouts.deco-0e@icloud.com>