Add Python script to check links in lopp.net
Recently, Jameson Lopp tweeted out that he had a link problem. He was
spending large amounts of his time checking all of the links he has on
his website for all of the Bitcoin resources. To help him I've been
working on this script to go through the website and fix all the links
that can't be accessed.