Fix bugs in lopp.net link checker
The lopp.net Python link checker will now filter out the links that are
correctly deemed valid by the validators package but actually just link
back to a specific section on the page. Not only can you not download
them but it's pretty safe to assume that they will exist.