Reorganize and Delete Miscellaneous Scripts
All of the Miscellaneous Scripts are now in a single folder instead of
separated into the Need Work and Finished sections. Several of my very
first script that aren't useful or interesting were pruned from the
repository. It felt a little sad to remove those scripts but I think it
was for the best. I've simply outgrown them, they don't do anything that
I think is worth keeping.