Add beginning of script to diagram sentences
In school I have to diagram sentences according to the Reed-Kellog
system. As I slaved away I began to wonder, am I really just applying
the same set of rules over and over to different sentences? And if those
rules are applied correctly than will the sentence be correctly
diagrammed? The answer seemed to be yes. If we accept this than it
naturally follows that all mistakes are simply from rules being
miss-applied by subjective and fallible humans.

But computers work by repetitively applying logical rules, they would be
the perfect sentence diagrammers. This is my experiment to see if I am
able to correctly break down and understand the rules of grammar enough
to explain them step by step to a computer.

So far I've set up a parts of speech tagger to create the base word
information I'll need to apply the rules and I've taught it to recognize
prepositional phrases. We'll see how it goes from here!