Fix hangman game and improve renaming script
I finally came back and fixed the hangman game I had tried to create
when I was first learning Python. It was laughable how I had tried to
approach the problem the first time. I must not have realized how much
easier it is to work with lists rather than strings.

The renaming script has now been improved to put leading zeroes in front
of filenames to make sure that they are always displayed in the correct
order.