Update grade calculator
The curve option in the grade calculator is much better now. If you pass
it the highest score in the class then it will calculate how much your
grade was adjusted by (assuming you didn't have the highest score). It
assumes the teacher was using the boost the highest score to 100%
method.